提现功能,未完善

+++ fanyin-service/src/main/java/com/fanyin/enums/RechargeStatus.java
@@ -25,19 +25,19 @@ public enum  RechargeStatus {
     /**
      * 状态
      */
-    private int code;
+    private byte code;
 
     /**
      * msg
      */
     private String msg;
 
-    RechargeStatus(int code, String msg) {
+    RechargeStatus(byte code, String msg) {
         this.code = code;
         this.msg = msg;
     }
 
-    public int getCode() {
+    public byte getCode() {
         return code;
     }
